ModuleClassLoaderTransformer extracted from JBossModulePlugin
